Run Aqua at all, and fix the three things it found
Aqua had never run. The gate read `VERSION == v"1.10"`, and VERSION is 1.10.12
while v"1.10" means v"1.10.0", so the condition was false on every version,
including the one it was written for -- the suite only ever paid for
precompiling Aqua. It now runs under `VERSION >= v"1.12"`, pinned rather than
unconditional because Aqua's findings (ambiguities above all) differ between
Julia versions, and 1.12 is what CI runs and what routine work targets.

What it found, all real:

- two exported symbols that do not exist, SearchResult and warmupbuild. The
  export was the bug; nothing could ever have used them.
- JSON declared as a dependency and never loaded -- src/ mentions it only in a
  log.jl docstring about emitting JSON, which this package does not do.
- three method ambiguities, all one cause: getindex(D::SparseDatabase, i) took
  an untyped i and so competed with getindex(db::AbstractDatabase, lst) for
  AbstractVector/AbstractVector{Bool}/BitArray. It indexes a column
  (nzrange(D.M, i)), so i::Integer is what it always meant; the vector forms now
  reach AbstractDatabase's SubDatabase methods as intended.

`[compat] julia` stays at 1.10+ and @BATCHES keeps its VERSION >= v"1.11" gates
around :greedy: nothing here needs a newer floor, and those gates are what keep
the declared floor honest. AGENTS.md now says so explicitly, along with how to
run an older version at all (the repo's Manifest.toml is resolved by whichever
Julia last touched it, so 1.10/1.11 need a throwaway environment) and the one
known failure there: test/testprojections.jl:167 fails 2 of 2 full-suite runs on
1.10.12, passes alone and on 1.12, unexplained.

Full suite green on 1.12, Aqua included (7 checks + zero ambiguities).

+++ src/special/sparse.jl
@@ -136,7 +136,7 @@ function LinearAlgebra.normalize!(A::SparseVecView)
     A
 end
 
-function Base.getindex(D::SparseDatabase, i)
+function Base.getindex(D::SparseDatabase, i::Integer)
     r = nzrange(D.M, i)
     rows = rowvals(D.M)
     vals = nonzeros(D.M)
